Refactor query logic and add to CLI
- Run queries via CLI command 'sonobuoy query'
- Modifies config to support overriding query output location so that
it works via CLI and via aggregator.
- Refactor the aggregator logic to utilize the new query organization
and not error the run if queries fail.
