+++ test/integration/sonobuoy_integration_test.go
@@ -342,6 +342,16 @@ func deleteComplete(t *testing.T, ns string) error {
 	return nil
 }
 
+func TestQuery(t *testing.T) {
+	t.Parallel()
+	ctx, cancel := context.WithTimeout(context.Background(), defaultTestTimeout)
+	defer cancel()
+
+	args := fmt.Sprint("query --level=trace")
+	output := mustRunSonobuoyCommandWithContext(ctx, t, "", args)
+	saveToArtifacts(t, output.String())
+}
+
 func TestConfigmaps(t *testing.T) {
 	t.Parallel()
 	ctx, cancel := context.WithTimeout(context.Background(), defaultTestTimeout)
@@ -429,7 +439,7 @@ func saveToArtifacts(t *testing.T, p string) (newPath string) {
 	var stdout, stderr bytes.Buffer
 
 	// Shell out to `mv` instead of using os.Rename(); the latter caused a problem due to files being on different devices.
-	cmd := exec.CommandContext(context.Background(), bash, "-c", fmt.Sprintf("mv %v %v", origFile, artifactFile))
+	cmd := exec.CommandContext(context.Background(), bash, "-c", fmt.Sprintf("mv -r %v %v", origFile, artifactFile))
 	cmd.Stdout = &stdout
 	cmd.Stderr = &stderr
